Add a Rakefile.
We propose a Rakefile as an alternative to the already provided scripts.
Ruby tends to be a little easier to reason about than shell scripts, and
rake adds a few niceties on top.

We have implemented all the previously available functionality. If
modified to use the new system, all tests pass.